requirements.txt
# Python 3.9+ recommended (3.12 OK). Before install: python -m pip install -U pip setuptools wheel
# Older pins (e.g. pandas<2.1) force source builds on 3.12 and often fail (e.g. missing pkg_resources).
pandas>=2.1.0,<3
scikit-learn>=1.3.0,<2
numpy>=1.26.0,<3
joblib>=1.3.0
streamlit>=1.28.0
openai>=1.30.0,<3
python-dotenv>=1.0.0,<2
matplotlib>=3.7.0,<4
